requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=8.0.0
scikit-learn>=1.0.0
mlflow>=2.0.0
tqdm>=4.65.0


pip install torch torchvision numpy Pillow scikit-learn mlflow
